Moved DiscovererComponent into appropriate place in Framework
